LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/offsets-into-the-media-resource/currentTime-move-within-document-expected.txt


FAIL playback should not reset when moving within a document assert_greater_than_equal: expected a number greater than or equal to 10 but got 0
